+++ plugins/input/csv/csv_index_featureset.hpp
@@ -2,7 +2,7 @@
  *
  * This file is part of Mapnik (c++ mapping toolkit)
  *
- * Copyright (C) 2016 Artem Pavlenko
+ * Copyright (C) 2017 Artem Pavlenko
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -40,7 +40,7 @@
 
 class csv_index_featureset : public mapnik::Featureset
 {
-    using value_type = std::pair<std::size_t, std::size_t>;
+    using value_type = std::pair<std::uint64_t, std::uint64_t>;
     using locator_type = csv_utils::geometry_column_locator;
 public:
